fixing type errors

+++ mgplvm/likelihoods.py
@@ -180,7 +180,7 @@ class Gaussian(Likelihood):
     @property
     def msg(self):
         sig = np.mean(self.sigma.data.cpu().numpy())
-        return (' lik_sig {:.3f} |').format(sig)
+        return (' lik_sig {:.3f} |').format(sig.astype(float))
 
 
 class Poisson(Likelihood):
@@ -451,4 +451,4 @@ class NegativeBinomial(Likelihood):
     @property
     def msg(self):
         total_count = np.mean(self.prms[0].data.cpu().numpy())
-        return (' lik_count {:.3f} |').format(total_count)
+        return (' lik_count {:.3f} |').format(total_count.astype(float))
